refactor: 팀-대회 소속 검증을 TeamConvenience로 이동
